@@ -18,40 +18,39 @@
   "nativescript": {
     "id": "org.nativescript.groceries",
     "tns-android": {
-      "version": "3.0.0"
+      "version": "3.4.0"
     },
     "tns-ios": {
-      "version": "3.0.0"
+      "version": "3.4.0"
     }
   },
   "scripts": {
     "tslint": "tslint -p tsconfig.json"
   },
   "dependencies": {
-    "@angular/animations": "~4.1.0",
-    "@angular/common": "~4.1.0",
-    "@angular/compiler": "~4.1.0",
-    "@angular/core": "~4.1.0",
-    "@angular/forms": "~4.1.0",
-    "@angular/http": "~4.1.0",
-    "@angular/platform-browser": "~4.1.0",
-    "@angular/platform-browser-dynamic": "~4.1.0",
-    "@angular/router": "~4.1.0",
+    "@angular/animations": "~5.2.0",
+    "@angular/common": "~5.2.0",
+    "@angular/compiler": "~5.2.0",
+    "@angular/core": "~5.2.0",
+    "@angular/forms": "~5.2.0",
+    "@angular/http": "~5.2.0",
+    "@angular/platform-browser": "~5.2.0",
+    "@angular/platform-browser-dynamic": "~5.2.0",
+    "@angular/router": "~5.2.0",
     "email-validator": "~1.0.7",
-    "nativescript-angular": "~3.0.0",
-    "nativescript-social-share": "~1.3.2",
+    "nativescript-angular": "~5.2.0",
+    "nativescript-social-share": "~1.4.0",
     "reflect-metadata": "~0.1.8",
-    "rxjs": "~5.3.0",
-    "tns-core-modules": "~3.0.0",
-    "zone.js": "~0.8.2"
+    "rxjs": "~5.5.0",
+    "tns-core-modules": "~3.4.0",
+    "zone.js": "~0.8.4"
   },
   "devDependencies": {
-    "babel-traverse": "6.25.0",
-    "babel-types": "6.25.0",
-    "babylon": "6.17.4",
+    "babel-traverse": "6.26.0",
+    "babel-types": "6.26.0",
+    "babylon": "6.18.0",
     "lazy": "1.0.11",
-    "nativescript-dev-typescript": "~0.4.0",
-    "typescript": "~2.2.1",
-    "tslint": "^5.4.2"
+    "nativescript-dev-typescript": "~0.6.0",
+    "typescript": "~2.6.2"
   }
 }
