@@ -1,26 +1,30 @@
 import { Injectable } from "@angular/core";
-import { Http, Headers } from "@angular/http";
+import { Http, Headers, URLSearchParams } from "@angular/http";
 import { Config } from "../config";
 import { Grocery } from "./grocery";
-import { Observable } from "rxjs/Rx";
+import { Observable } from "rxjs/Observable";
+import "rxjs/add/operator/catch";
 import "rxjs/add/operator/map";
 
 @Injectable()
 export class GroceryListService {
+  baseUrl = Config.apiUrl + "appdata/" + Config.appKey + "/Groceries";
   constructor(private http: Http) {}
 
   load() {
-    let headers = new Headers();
-    headers.append("Authorization", "Bearer " + Config.token);
+    // Kinvey-specific syntax to sort the groceries by last modified time. Don’t worry about the details here.
+    let params = new URLSearchParams();
+    params.append("sort", "{\"_kmd.lmt\": 1}");
 
-    return this.http.get(Config.apiUrl + "Groceries", {
-      headers: headers
+    return this.http.get(this.baseUrl, {
+      headers: this.getCommonHeaders(),
+      params: params
     })
     .map(res => res.json())
     .map(data => {
       let groceryList = [];
-      data.Result.forEach((grocery) => {
-        groceryList.push(new Grocery(grocery.Id, grocery.Name));
+      data.forEach((grocery) => {
+        groceryList.push(new Grocery(grocery._id, grocery.Name));
       });
       return groceryList;
     })
@@ -28,33 +32,32 @@ export class GroceryListService {
   }
 
   add(name: string) {
-    let headers = new Headers();
-    headers.append("Authorization", "Bearer " + Config.token);
-    headers.append("Content-Type", "application/json");
-
     return this.http.post(
-      Config.apiUrl + "Groceries",
+      this.baseUrl,
       JSON.stringify({ Name: name }),
-      { headers: headers }
+      { headers: this.getCommonHeaders() }
     )
     .map(res => res.json())
     .map(data => {
-      return new Grocery(data.Result.Id, name);
+      return new Grocery(data._id, name);
     })
     .catch(this.handleErrors);
   }
 
   delete(id: string) {
-    let headers = new Headers();
-    headers.append("Authorization", "Bearer " + Config.token);
-    headers.append("Content-Type", "application/json");
-
     return this.http.delete(
-      Config.apiUrl + "Groceries/" + id,
-      { headers: headers }
+      this.baseUrl + "/" + id,
+      { headers: this.getCommonHeaders() }
     )
     .map(res => res.json())
     .catch(this.handleErrors);
+  }
+
+  getCommonHeaders() {
+    let headers = new Headers();
+    headers.append("Content-Type", "application/json");
+    headers.append("Authorization", "Kinvey " + Config.token);
+    return headers;
   }
 
   handleErrors(error: Response) {
