@@ -1,6 +1,7 @@
 import { Injectable } from "@angular/core";
 import { Http, Headers, Response } from "@angular/http";
-import { Observable } from "rxjs/Rx";
+import { Observable } from "rxjs/Observable";
+import "rxjs/add/operator/catch";
 import "rxjs/add/operator/do";
 import "rxjs/add/operator/map";
 
@@ -12,39 +13,40 @@ export class UserService {
   constructor(private http: Http) {}
 
   register(user: User) {
-    let headers = new Headers();
-    headers.append("Content-Type", "application/json");
-
     return this.http.post(
-      Config.apiUrl + "Users",
+      Config.apiUrl + "user/" + Config.appKey,
       JSON.stringify({
-        Username: user.email,
-        Email: user.email,
-        Password: user.password
+        username: user.email,
+        email: user.email,
+        password: user.password
       }),
-      { headers: headers }
+      { headers: this.getCommonHeaders() }
     )
     .catch(this.handleErrors);
   }
 
   login(user: User) {
-    let headers = new Headers();
-    headers.append("Content-Type", "application/json");
-
     return this.http.post(
-      Config.apiUrl + "oauth/token",
+      Config.apiUrl + "user/" + Config.appKey + "/login",
       JSON.stringify({
         username: user.email,
         password: user.password,
         grant_type: "password"
       }),
-      { headers: headers }
+      { headers: this.getCommonHeaders() }
     )
     .map(response => response.json())
     .do(data => {
-      Config.token = data.Result.access_token;
+      Config.token = data._kmd.authtoken;
     })
     .catch(this.handleErrors);
+  }
+
+  getCommonHeaders() {
+    let headers = new Headers();
+    headers.append("Content-Type", "application/json");
+    headers.append("Authorization", Config.authHeader);
+    return headers;
   }
 
   handleErrors(error: Response) {
